+++ tests/test_gail.py
@@ -8,7 +8,9 @@ from stable_baselines import A2C, ACER, ACKTR, GAIL, DDPG, DQN, PPO1, PPO2,\
  TD3, TRPO, SAC
 from stable_baselines.common.cmd_util import make_atari_env
 from stable_baselines.common.vec_env import VecFrameStack
+from stable_baselines.common.evaluation import evaluate_policy
 from stable_baselines.gail import ExpertDataset, generate_expert_traj
+
 
 EXPERT_PATH_PENDULUM = "stable_baselines/gail/dataset/expert_pendulum.npz"
 EXPERT_PATH_DISCRETE = "stable_baselines/gail/dataset/expert_cartpole.npz"
@@ -36,13 +38,7 @@ def test_gail(expert_env):
     model = model.load("GAIL-{}".format(env_id), env=env)
     model.learn(1000)
 
-    obs = env.reset()
-
-    for _ in range(1000):
-        action, _ = model.predict(obs)
-        obs, _, done, _ = env.step(action)
-        if done:
-            obs = env.reset()
+    evaluate_policy(model, env, n_eval_episodes=5)
     del dataset, model
 
 @pytest.mark.parametrize("generate_env", [