+++ tests/test_save.py
@@ -7,12 +7,12 @@ import pytest
 import numpy as np
 
 from stable_baselines import A2C, ACER, ACKTR, DQN, PPO1, PPO2, TRPO
-from stable_baselines.common import set_global_seeds
 from stable_baselines.common.identity_env import IdentityEnv
 from stable_baselines.common.vec_env import DummyVecEnv
+from stable_baselines.common.evaluation import evaluate_policy
 from stable_baselines.common.policies import MlpPolicy, FeedForwardPolicy
 
-N_TRIALS = 2000
+N_EVAL_EPISODES = 100
 
 MODEL_LIST = [
     A2C,
@@ -44,11 +44,11 @@ def test_model_manipulation(request, model_class, storage_method, store_format):
     works and that the action prediction works
 
     :param model_class: (BaseRLModel) A RL model
-    :param storage_method: (str) Should file be saved to a file ("path") or to a buffer 
+    :param storage_method: (str) Should file be saved to a file ("path") or to a buffer
         ("file-like")
     :param store_format: (str) Save format, either "zip" or "cloudpickle".
     """
-    
+
     # Use postfix ".model" so we can remove the file later
     model_fname = './test_model_{}.model'.format(request.node.name)
     store_as_cloudpickle = store_format == "cloudpickle"
@@ -57,20 +57,12 @@ def test_model_manipulation(request, model_class, storage_method, store_format):
         env = DummyVecEnv([lambda: IdentityEnv(10)])
 
         # create and train
-        model = model_class(policy="MlpPolicy", env=env)
-        model.learn(total_timesteps=50000)
+        model = model_class(policy="MlpPolicy", env=env, seed=0)
+        model.learn(total_timesteps=10000)
 
-        # predict and measure the acc reward
-        acc_reward = 0
-        set_global_seeds(0)
-        obs = env.reset()
-        for _ in range(N_TRIALS):
-            action, _ = model.predict(obs)
-            # Test action probability method
-            model.action_probability(obs)
-            obs, reward, _, _ = env.step(action)
-            acc_reward += reward
-        acc_reward = sum(acc_reward) / N_TRIALS
+        env.envs[0].action_space.seed(0)
+        mean_reward, _ = evaluate_policy(model, env, deterministic=True,
+                                         n_eval_episodes=N_EVAL_EPISODES)
 
         # test action probability for given (obs, action) pair
         env = model.get_env()
@@ -107,37 +99,24 @@ def test_model_manipulation(request, model_class, storage_method, store_format):
         model.set_env(env)
 
         # predict the same output before saving
-        loaded_acc_reward = 0
-        set_global_seeds(0)
-        obs = env.reset()
-        for _ in range(N_TRIALS):
-            action, _ = model.predict(obs)
-            obs, reward, _, _ = env.step(action)
-            loaded_acc_reward += reward
-        loaded_acc_reward = sum(loaded_acc_reward) / N_TRIALS
-        assert abs(acc_reward - loaded_acc_reward) < 0.1, "Error: the prediction seems to have changed between " \
-                                                          "loading and saving"
+        env.envs[0].action_space.seed(0)
+        loaded_mean_reward, _ = evaluate_policy(model, env, deterministic=True, n_eval_episodes=N_EVAL_EPISODES)
+        # Allow 10% diff
+        assert abs((mean_reward - loaded_mean_reward) / mean_reward) < 0.1, "Error: the prediction seems to have changed between " \
+                                                                            "loading and saving"
 
         # learn post loading
         model.learn(total_timesteps=100)
 
         # validate no reset post learning
-        loaded_acc_reward = 0
-        set_global_seeds(0)
-        obs = env.reset()
-        for _ in range(N_TRIALS):
-            action, _ = model.predict(obs)
-            obs, reward, _, _ = env.step(action)
-            loaded_acc_reward += reward
-        loaded_acc_reward = sum(loaded_acc_reward) / N_TRIALS
-        assert abs(acc_reward - loaded_acc_reward) < 0.1, "Error: the prediction seems to have changed between " \
-                                                          "pre learning and post learning"
+        env.envs[0].action_space.seed(0)
+        loaded_mean_reward, _ = evaluate_policy(model, env, deterministic=True, n_eval_episodes=N_EVAL_EPISODES)
+
+        assert abs((mean_reward - loaded_mean_reward) / mean_reward) < 0.15, "Error: the prediction seems to have changed between " \
+                                                                            "pre learning and post learning"
 
         # predict new values
-        obs = env.reset()
-        for _ in range(N_TRIALS):
-            action, _ = model.predict(obs)
-            obs, _, _, _ = env.step(action)
+        evaluate_policy(model, env, n_eval_episodes=N_EVAL_EPISODES)
 
         del model, env